Minor fixes to make armv7hl build as arm-efi
- /remove/ the code to use use softfloat unconditionally on arm
  (I can't get our arm bins to build with it for whatever reason...)
- Fix one type error
- build the secure boot loader code on arm

+++ grub-core/loader/efi/linux.c
@@ -50,7 +50,7 @@ grub_linuxefi_secure_validate (void *data, grub_uint32_t size)
 
   grub_dprintf ("secureboot", "Asking shim to verify kernel signature\n");
   status = shim_lock->verify (data, size);
-  grub_dprintf ("secureboot", "shim_lock->verify(): %ld\n", status);
+  grub_dprintf ("secureboot", "shim_lock->verify(): %ld\n", (long int)status);
   if (status == GRUB_EFI_SUCCESS)
     {
       grub_dprintf ("secureboot", "Kernel signature verification passed\n");
